chore(repo): Rename `test/utils` to `test/test_utils` to prevent confusion.
